Bugfix for issue #2
The child count of a node can increase while parsing it if we find
numbers and convert them into dial tooltips. If all the numbers are
childs of one node and we iterate over it then we have to read the
length of the child nodes for each loop otherwise the parser will stop
to early.